Building out the API.
Generally speaking, I am pretty happy now with the structure of the
app and its tests. This is not to say the tests are correct or
complete, just that they seem to be coming together in a way that
makes sense given the authentication needs of the API.

Some notes:
* The user tests are in better shape than the sources tests.
* The /status route may or may not be a good idea. The intention
  is to make available an action to help a client check on its
  status with the API.

Note that this a squashed commit from development this week.