Big refactoring ... using Sinatra::Base now
Using Sinatra::Base now instead of a monolithic Sinatra app.

This means the routes in the code and tests are shorter. The
various sub-apps are hooked together in the rackup file:
config.ru.

Extracted out common patterns in controllers. Now using a DSL
to make it easier to create new resource. This DSL includes
some callbacks (more probably will be added in the future).
Lots of code was eliminated.

To elaborate:
* controller-level validations are more standardized
* permissions checking is cleaned up
* created a common pattern for nested resources

Also, controller tests are much more standardized. The intent
is that you can cut-and-paste from existing tests as a good
starting point.

Various:
* We use the DataCatalog namespace now. Perhaps more can be
  done here, too?
* Several code chunks of to_json were removed since they were
  no longer needed.