Added aggregate ratings to Comment and Source
The original story that prompted this change is here:
http://www.pivotaltracker.com/story/show/1132776

I decided a good way to implement this feature was to use observers.
So I added an observer to Rating. Note that observers live in their
own directory, which is loaded after all the models have loaded.

Since we expect reads (of an aggregate) to be more common than writes,
we decided to make the writes a little bit expensive and the reads
cheap.

I also had to fix up MongoMapper proxies in order for #nil? to work
on MongoMapper associations. See:
http://github.com/djsun/mongomapper/tree/add-nil-to-proxies

and, to make debugging easier:
http://github.com/djsun/mongomapper/tree/add-inspect-to-proxies

Hopefully these patches will get pulled into MongoMapper soon. For
the time being, I am again relying on my own version of the gem:
djsun-mongomapper version 0.3.5.1.