Added tests for permission helpers
(These are 'almost' unit tests, but not quite.)

The way I stubbed out 'error' could use improvement. The behavior
of 'error' in Sinatra is that execution of an action is halted and
a page is returned with the error code and message.

I didn't really model this behavior accurately in the stub for
'error'. It only 'records' the fact that it was called, but does not
cause an exception to be raised.

Therefore, I needed to add 'return' statements to 'require_at_least'
in permission_helpers.rb to make the tests pass.