Split out /resources from /checkup
Relaxed requirement that GET / yell if there was an API key. That was
silly. The reason probably had to do with showing the appropriate
resources available based on a given API key. Now that /resources is
separate from /checkup and / this should not be a problem.

Note: renamed root to _root just so that an alphabetical listing looks
better (e.g. shows root -- which is really "/" -- at the top.)