409 Conflict on attempt to delete primary key
Deleting a primary key would put the user resource in an inconsistent
state, since all users must always have a primary key.

Previously, I had used 403 Forbidden, but it was not a good fit.
According to RESTful Web Services, 403 is used when for "a resource
that is only accessible at certain times, or from certain IP
addresses." Put another way, "the server made its decision based
on some aspect of the request other that its form; say, the time of
day the request was made."